backend/requirements.txt
Django==2.2.1
django-cors-headers==3.0.1
django-rest-auth==0.9.5
djangorestframework==3.9.4
pytz==2019.1
six==1.12.0
sqlparse==0.3.0